% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pies.R
\name{add_pies}
\alias{add_pies}
\title{add_pies}
\usage{
add_pies(map, locations, pies, sizes)
}
\arguments{
\item{map}{ggplot map}

\item{locations}{location data.frame}

\item{pies}{a list of pies}

\item{sizes}{size of pie to be added}
}
\value{
ggplot2 object
}
\description{
add a list of pies to map
}
\author{
Guangchuang Yu
}

