% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pies.R
\name{add_pies2}
\alias{add_pies2}
\title{add_pies2}
\usage{
add_pies2(map, data, cols, color, alpha = 1, scale_fun = function(x)
  x/sum(x))
}
\arguments{
\item{map}{map}

\item{data}{data contains column of place, longitude, latitude and other columns for drawing pie}

\item{cols}{columns of stats for drawing pie}

\item{color}{color of pie}

\item{alpha}{transpancy of pie}

\item{scale_fun}{function to scale pie size}
}
\value{
ggplot2 object
}
\description{
add pies to a map by user provided data
}
\author{
Guangchuang Yu
}

