% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{draw_map}
\alias{draw_map}
\title{draw_map}
\usage{
draw_map(database = "world", regions = ".", fill = NA, color = "grey50",
  xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{database}{name of map provided by 'maps' package.  These include
'county', 'france', 'italy', 'nz', 'state', 'usa', 'world',
'world2'.}

\item{regions}{name of subregions to include. Defaults to '.' which includes all subregion.
See documentation for 'map' for more details}

\item{fill}{color to fill}

\item{color}{border color}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{...}{additional parameter}
}
\value{
ggplot2 object
}
\description{
draw a map
}
\author{
Guangchuang Yu
}

