% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map_by_country}
\alias{get_map_by_country}
\title{get_map_by_country}
\usage{
get_map_by_country(country)
}
\arguments{
\item{country}{selected country or countries}
}
\value{
map data in data.frame
}
\description{
extract map data of selected country from 'world' map
}
\author{
ygc
}

