% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map_by_region}
\alias{get_map_by_region}
\title{get_map_by_region}
\usage{
get_map_by_region(region)
}
\arguments{
\item{region}{selected region(s)}
}
\value{
map data in data.frame
}
\description{
extract map data of selected region(s) from 'world' map
}
\author{
ygc
}

