% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map_china}
\alias{get_map_china}
\title{get_map_data}
\usage{
get_map_china(province = "all")
}
\arguments{
\item{province}{selected province(s) or 'all'}
}
\value{
map data of China, in data.frame
}
\description{
get map data of China or province(s) of China
}
\author{
ygc
}

