% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pies.R
\name{pie2}
\alias{pie2}
\title{pie2}
\usage{
pie2(data, y, fill, color, alpha = 1)
}
\arguments{
\item{data}{data.frame}

\item{y}{y variable}

\item{fill}{color to fill}

\item{color}{border color}

\item{alpha}{transparency}
}
\value{
ggplot2 object
}
\description{
pie plot by ggplot2
}
\author{
Guangchuang Yu
}

